backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
praw>=7.7.0
youtube-transcript-api>=0.6.0
google-genai>=0.3.0
google-api-python-client>=2.100.0
cachetools>=5.3.0
yfpy>=15.0.0
